Removed Opcode abstraction
The enum abstraction for the different opcodes was less than
satisfactory because:
+ It hid the "true name" of the opcodes behind something easier to
remember, but it was very unhelpful when looking at manuals.
+ It was clunky. The enum was way too long, and a lot of code was just
mirroring the match expression when constructing opcodes.
+ It didn't offer enough. Just being an alias for opcodes shouldn't be
enough. I would have kept it if, e.g. i wanted to do some
transformations or interesting operations with opcodes, but I didn't
need it.